requirements.txt
# FastAPI and server
fastapi>=0.100.0
uvicorn[standard]>=0.20.0

# Chess library for encoding/decoding
python-chess>=1.999

# Authentication and security
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4

# Database
sqlalchemy>=2.0.0
mysql-connector-python>=8.0.33

# Utilities
werkzeug>=2.3.0
